Made changes for the cart total and superscript
